增加了从csv文件读取数据功能，可以通过该方法注入数据。 其他修改：Transaction的成员变量gasPrice的长度超过了long，因此修改为double类型

+++ ArraySimulator/src/shardSystem/shardNode.java
@@ -173,7 +173,7 @@ public class shardNode extends Replica {
 			pstmt.setString(2, tx.getRecipient());
 			pstmt.setDouble(3, tx.getValue());
 			pstmt.setLong(4, tx.getTimestamp());
-			pstmt.setLong(5, tx.getGasPrice());
+			pstmt.setDouble(5, tx.getGasPrice());
 			pstmt.setLong(6, tx.getAccountNonce());
 
             pstmt.executeUpdate();
